+++ examples/sentry-metrics/runner.py
@@ -6,7 +6,6 @@
 
 import argparse
 import os
-from typing import Optional
 
 import aiohttp
 
@@ -14,15 +13,7 @@ from pipecat.transports.services.helpers.daily_rest import DailyRESTHelper
 
 
 async def configure(aiohttp_session: aiohttp.ClientSession):
-    (url, token, _) = await configure_with_args(aiohttp_session)
-    return (url, token)
-
-
-async def configure_with_args(
-    aiohttp_session: aiohttp.ClientSession, parser: Optional[argparse.ArgumentParser] = None
-):
-    if not parser:
-        parser = argparse.ArgumentParser(description="Daily AI SDK Bot Sample")
+    parser = argparse.ArgumentParser(description="Daily AI SDK Bot Sample")
     parser.add_argument(
         "-u", "--url", type=str, required=False, help="URL of the Daily room to join"
     )
@@ -61,4 +52,5 @@ async def configure_with_args(
 
     token = await daily_rest_helper.get_token(url, expiry_time)
 
-    return (url, token, args)
+    return (url, token)
+    return (url, token)